+++ app/src/main/kotlin/hr/eduwalk/ui/event/EditLocationEvent.kt
@@ -2,5 +2,5 @@ package hr.eduwalk.ui.event
 
 sealed interface EditLocationEvent {
 
-    data class FinishDeleteLocation(val locationId: Int) : EditLocationEvent
+    data class FinishDeleteLocation(val locationId: Long) : EditLocationEvent
 }
